refactor(arch): Phase 4 — PR #485 리뷰 2차 수정
- B1: byokGate.ts @/entities/analysis → ./types (shared 내부)
- B2: llmProviderLabels.ts @/entities/api-key → ../config/llmProviders
- B3: contactErrorMessages.ts → shared/config/contact.ts 신규, 상수 이동
- B4: skillStats.ts → shared/lib/skillUtils.ts 신규, countSkillsByType 이동
- S1: OptionsExpirationSelector JSDoc WHY 복원
