+++ src/entities/inquiry/lib/constants.ts
@@ -1,3 +1,5 @@
-export const CONTACT_TITLE_MAX_LENGTH = 120;
-export const CONTACT_CONTENT_MAX_LENGTH = 2000;
+export {
+    CONTACT_CONTENT_MAX_LENGTH,
+    CONTACT_TITLE_MAX_LENGTH,
+} from '@/shared/config/contact';
 export const CONTACT_EMAIL_PATTERN = /^[^\s@]+@[^\s@]+\.[^\s@]+$/u;
