+++ src/entities/skill/lib/skills.ts
@@ -1,10 +1,1 @@
-import type { SkillShowcaseItem, SkillType } from '@y0ngha/siglens-core';
-
-export function countSkillsByType(
-    skills: readonly SkillShowcaseItem[]
-): Partial<Record<SkillType, number>> {
-    return skills.reduce<Partial<Record<SkillType, number>>>((acc, skill) => {
-        if (skill.type == null) return acc;
-        return { ...acc, [skill.type]: (acc[skill.type] ?? 0) + 1 };
-    }, {});
-}
+export { countSkillsByType } from '@/shared/lib/skillUtils';
