+++ src/shared/config/contact.ts
@@ -1,0 +1,2 @@
+export const CONTACT_TITLE_MAX_LENGTH = 120;
+export const CONTACT_CONTENT_MAX_LENGTH = 2000;
