+++ src/shared/lib/types.ts
@@ -137,9 +137,13 @@ export interface AnalysisGateBlockedResult {
 }
 
 /**
- * UI-level expiration filter value used by the expiration selector and
- * downstream chain/metrics views. An ISO 'YYYY-MM-DD' string selects a
- * single expiration; `'all'` aggregates every expiration in the snapshot.
+ * UI-level expiration filter value.
+ *
+ * The `(string & {})` intersection prevents TypeScript from widening the
+ * union to bare `string` (which would drop the `'all'` autocomplete in
+ * IDEs and erase the literal hint at call sites). Runtime behavior is
+ * identical to `string | 'all'`; this trick is purely an editor-DX guard.
+ * Mirrors siglens-core's OptionsExpirationFilter for consistency.
  */
 export type OptionsExpirationSelector = (string & {}) | 'all';
